# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.1.0"
  title: express-swagger-rest
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /swagger:
    x-swagger-pipe: swagger_raw
  /comics:
    x-swagger-router-controller: comic
    get:
      description: Returns all comics
      operationId: comic_list
      responses:
        200:
          description: JSON array of all comics
          schema:
            type: array
            items:
              $ref: "#/definitions/Comic"
        500:
          description: Server error
          schema:
            type: object
    post:
      description: Create a new comic
      operationId: comic_create
      parameters:
      - name: comic
        in: body
        description: JSON representation of the comic to be created
        required: true
        schema:
          $ref: "#/definitions/Comic"
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/Comic"
        400:
          description: Bad request
          schema:
            type: object
        500:
          description: Server error
          schema:
            type: object
  /comics/bulk:
    x-swagger-router-controller: comic
    post:
      description: Create a series of new comics
      operationId: comic_bulkCreate
      parameters:
      - name: comics
        in: body
        description: JSON representation of the comic to be created
        required: true
        schema:
          type: array
          items:
            $ref: "#/definitions/Comic"
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/Comic"
        400:
          description: Bad request
          schema:
            type: object
        500:
          description: Server error
          schema:
            type: object
  /comics/{id}:
    x-swagger-router-controller: comic
    get:
      description: Returns the comic with the requested ID
      operationId: comic_get
      parameters:
      - name: id
        in: path
        description: ID of the comic
        required: true
        type: integer
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Comic"
        404:
          description: Not found
          schema:
            type: object
        500:
          description: Server error
          schema:
            type: object
    put:
      description: Make changes to the given comic
      operationId: comic_put
      parameters:
      - name: id
        in: path
        description: ID of the comic
        required: true
        type: integer
      - name: data
        in: body
        description: Fields to update
        required: true
        schema:
          $ref: "#/definitions/Comic"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Comic"
        404:
          description: Not found
          schema:
            type: object
        500:
          description: Server error
          schema:
            type: object
  /comics/{comic}/page/{page}:
    x-swagger-router-controller: image
    get:
      description: Get the page of comics
      operationId: image_list
      parameters:
      - name: comic
        in: path
        description: ID of the comic
        required: true
        type: integer
      - name: page
        in: path
        description: Page number to fetch
        required: true
        type: integer
      - name: size
        in: query
        required: false
        description: Number of comics to retun on the page
        default: 25
        type: integer
      responses:
        200:
          description: Success
          schema:
            properties:
              count:
                type: integer
              rows:
                type: array
                items:
                  $ref: "#/definitions/Comic"
        404:
          description: Not found
          schema:
            type: object
        500:
          description: Server error
          schema:
            type: object

# complex objects have schema definitions
definitions:
  Comic:
    properties:
      id:
        type: integer
      title:
        type: string
      folder:
        type: string
      author:
        type: string
      description:
        type: string
      start:
        type: string
      latest:
        type: string
      linkSelector:
        type: string
      imgSelector:
        type: string
      noteSelector:
        type: string